Moved request, notification, hfclk responsibility out of nrf_802154_swi.c file
Using new module nrf_802154_queue
