+++ src/nrf_802154_request_swi.c
@@ -45,12 +45,165 @@
 #include "nrf_802154_critical_section.h"
 #include "nrf_802154_debug.h"
 #include "nrf_802154_peripherals.h"
+#include "nrf_802154_queue.h"
 #include "nrf_802154_rx_buffer.h"
 #include "nrf_802154_swi.h"
 #include "nrf_802154_utils.h"
 #include "nrf_radio.h"
+#include "nrf_egu.h"
 
 #include <nrf.h>
+
+#define SWI_EGU            NRF_802154_SWI_EGU_INSTANCE ///< Label of SWI peripheral.
+
+/** Size of requests queue.
+ *
+ * Two is minimal queue size. It is not expected in current implementation to queue a few requests.
+ */
+#define REQ_QUEUE_SIZE     2
+
+#define REQ_INT            NRF_EGU_INT_TRIGGERED2      ///< Label of request interrupt.
+#define REQ_TASK           NRF_EGU_TASK_TRIGGER2       ///< Label of request task.
+#define REQ_EVENT          NRF_EGU_EVENT_TRIGGERED2    ///< Label of request event.
+
+/// Type of requests in request queue.
+typedef enum
+{
+    REQ_TYPE_SLEEP,
+    REQ_TYPE_RECEIVE,
+    REQ_TYPE_TRANSMIT,
+    REQ_TYPE_ENERGY_DETECTION,
+    REQ_TYPE_CCA,
+    REQ_TYPE_CONTINUOUS_CARRIER,
+    REQ_TYPE_BUFFER_FREE,
+    REQ_TYPE_CHANNEL_UPDATE,
+    REQ_TYPE_CCA_CFG_UPDATE,
+    REQ_TYPE_RSSI_MEASURE,
+    REQ_TYPE_RSSI_GET,
+} nrf_802154_req_type_t;
+
+/// Request data in request queue.
+typedef struct
+{
+    nrf_802154_req_type_t type; ///< Type of the request.
+
+    union
+    {
+        struct
+        {
+            nrf_802154_term_t term_lvl; ///< Request priority.
+            bool            * p_result; ///< Sleep request result.
+        } sleep;                        ///< Sleep request details.
+
+        struct
+        {
+            nrf_802154_notification_func_t notif_func;  ///< Error notified in case of success.
+            nrf_802154_term_t              term_lvl;    ///< Request priority.
+            req_originator_t               req_orig;    ///< Request originator.
+            bool                           notif_abort; ///< If function termination should be notified.
+            bool                         * p_result;    ///< Receive request result.
+        } receive;                                      ///< Receive request details.
+
+        struct
+        {
+            nrf_802154_notification_func_t notif_func; ///< Error notified in case of success.
+            nrf_802154_term_t              term_lvl;   ///< Request priority.
+            req_originator_t               req_orig;   ///< Request originator.
+            const uint8_t                * p_data;     ///< Pointer to a buffer containing PHR and PSDU of the frame to transmit.
+            bool                           cca;        ///< If CCA was requested prior to transmission.
+            bool                           immediate;  ///< If TX procedure must be performed immediately.
+            bool                         * p_result;   ///< Transmit request result.
+        } transmit;                                    ///< Transmit request details.
+
+        struct
+        {
+            nrf_802154_term_t term_lvl; ///< Request priority.
+            bool            * p_result; ///< Energy detection request result.
+            uint32_t          time_us;  ///< Requested time of energy detection procedure.
+        } energy_detection;             ///< Energy detection request details.
+
+        struct
+        {
+            nrf_802154_term_t term_lvl; ///< Request priority.
+            bool            * p_result; ///< CCA request result.
+        } cca;                          ///< CCA request details.
+
+        struct
+        {
+            nrf_802154_term_t term_lvl; ///< Request priority.
+            bool            * p_result; ///< Continuous carrier request result.
+        } continuous_carrier;           ///< Continuous carrier request details.
+
+        struct
+        {
+            uint8_t * p_data;   ///< Pointer to receive buffer to free.
+            bool    * p_result; ///< Buffer free request result.
+        } buffer_free;          ///< Buffer free request details.
+
+        struct
+        {
+            bool * p_result; ///< Channel update request result.
+        } channel_update;    ///< Channel update request details.
+
+        struct
+        {
+            bool * p_result; ///< CCA config update request result.
+        } cca_cfg_update;    ///< CCA config update request details.
+
+        struct
+        {
+            bool * p_result; ///< RSSI measurement request result.
+        } rssi_measure;      ///< RSSI measurement request details.
+
+        struct
+        {
+            int8_t * p_rssi;                              ///< RSSI measurement result.
+            bool   * p_result;                            ///< RSSI measurement status.
+        } rssi_get;                                       ///< Details of the getter that retrieves the RSSI measurement result.
+    } data;                                               ///< Request data depending on its type.
+} nrf_802154_req_data_t;
+
+/**@brief Instance of a requests queue */
+static nrf_802154_queue_t    m_requests_queue;
+
+/**@brief Memory holding requests queue items */
+static nrf_802154_req_data_t m_requests_queue_memory[REQ_QUEUE_SIZE];
+
+/**
+ * Enter request block.
+ *
+ * This is a helper function used in all request functions to atomically
+ * find an empty slot in request queue and allow atomic slot update.
+ *
+ * @return Pointer to an empty slot in the request queue.
+ */
+static nrf_802154_req_data_t * req_enter(void)
+{
+    __disable_irq();
+    __DSB();
+    __ISB();
+
+    assert(!nrf_802154_queue_is_full(&m_requests_queue));
+
+    return (nrf_802154_req_data_t *)nrf_802154_queue_push_begin(&m_requests_queue);
+}
+
+/**
+ * Exit request block.
+ *
+ * This is a helper function used in all request functions to end atomic slot update
+ * and trigger SWI to process the request from the slot.
+ */
+static void req_exit(void)
+{
+    nrf_802154_queue_push_commit(&m_requests_queue);
+
+    nrf_egu_task_trigger(SWI_EGU, REQ_TASK);
+
+    __enable_irq();
+    __DSB();
+    __ISB();
+}
 
 /** Assert if SWI interrupt is disabled. */
 static inline void assert_interrupt_status(void)
@@ -99,14 +252,226 @@ static bool active_vector_priority_is_high(void)
     return nrf_802154_critical_section_active_vector_priority_get() <= NRF_802154_SWI_PRIORITY;
 }
 
+/**
+ * @brief Requests entering the @ref RADIO_STATE_SLEEP state from the SWI priority.
+ *
+ * @param[in]   term_lvl  Termination level of this request. Selects procedures to abort.
+ * @param[out]  p_result  Result of entering the sleep state.
+ */
+static void swi_sleep(nrf_802154_term_t term_lvl, bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                = REQ_TYPE_SLEEP;
+    p_slot->data.sleep.term_lvl = term_lvl;
+    p_slot->data.sleep.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Requests entering the @ref RADIO_STATE_RX state from the SWI priority.
+ *
+ * @param[in]   term_lvl         Termination level of this request. Selects procedures to abort.
+ * @param[in]   req_orig         Module that originates this request.
+ * @param[in]   notify_function  Function called to notify the status of the procedure. May be NULL.
+ * @param[in]   notify_abort     If abort notification should be triggered automatically.
+ * @param[out]  p_result         Result of entering the receive state.
+ */
+static void swi_receive(nrf_802154_term_t              term_lvl,
+                            req_originator_t               req_orig,
+                            nrf_802154_notification_func_t notify_function,
+                            bool                           notify_abort,
+                            bool                         * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                     = REQ_TYPE_RECEIVE;
+    p_slot->data.receive.term_lvl    = term_lvl;
+    p_slot->data.receive.req_orig    = req_orig;
+    p_slot->data.receive.notif_func  = notify_function;
+    p_slot->data.receive.notif_abort = notify_abort;
+    p_slot->data.receive.p_result    = p_result;
+
+    req_exit();
+}
+
+/**
+ * @biref Requests entering the @ref RADIO_STATE_TX state from the SWI priority.
+ *
+ * @param[in]   term_lvl         Termination level of this request. Selects procedures to abort.
+ * @param[in]   req_orig         Module that originates this request.
+ * @param[in]   p_data           Pointer to a buffer that contains PHR and PSDU of the frame to be
+ *                               transmitted.
+ * @param[in]   cca              If the driver should perform the CCA procedure before transmission.
+ * @param[in]   immediate        If true, the driver schedules transmission immediately or never;
+ *                               if false, the transmission may be postponed until TX preconditions
+ *                               are met.
+ * @param[in]   notify_function  Function called to notify the status of this procedure instead of
+ *                               the default notification. If NULL, the default notification
+ *                               is used.
+ * @param[out]  p_result         Result of entering the transmit state.
+ */
+static void swi_transmit(nrf_802154_term_t              term_lvl,
+                             req_originator_t               req_orig,
+                             const uint8_t                * p_data,
+                             bool                           cca,
+                             bool                           immediate,
+                             nrf_802154_notification_func_t notify_function,
+                             bool                         * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                     = REQ_TYPE_TRANSMIT;
+    p_slot->data.transmit.term_lvl   = term_lvl;
+    p_slot->data.transmit.req_orig   = req_orig;
+    p_slot->data.transmit.p_data     = p_data;
+    p_slot->data.transmit.cca        = cca;
+    p_slot->data.transmit.immediate  = immediate;
+    p_slot->data.transmit.notif_func = notify_function;
+    p_slot->data.transmit.p_result   = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Requests entering the @ref RADIO_STATE_ED state from the SWI priority.
+ *
+ * @param[in]   term_lvl  Termination level of this request. Selects procedures to abort.
+ * @param[in]   time_us   Requested duration of the energy detection procedure.
+ * @param[out]  p_result  Result of entering the energy detection state.
+ */
+static void swi_energy_detection(nrf_802154_term_t term_lvl,
+                                     uint32_t          time_us,
+                                     bool            * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                           = REQ_TYPE_ENERGY_DETECTION;
+    p_slot->data.energy_detection.term_lvl = term_lvl;
+    p_slot->data.energy_detection.time_us  = time_us;
+    p_slot->data.energy_detection.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Requests entering the @ref RADIO_STATE_CCA state from the SWI priority.
+ *
+ * @param[in]   term_lvl  Termination level of this request. Selects procedures to abort.
+ * @param[out]  p_result  Result of entering the CCA state.
+ */
+static void swi_cca(nrf_802154_term_t term_lvl, bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type              = REQ_TYPE_CCA;
+    p_slot->data.cca.term_lvl = term_lvl;
+    p_slot->data.cca.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Requests entering the @ref RADIO_STATE_CONTINUOUS_CARRIER state from the SWI priority.
+ *
+ * @param[in]   term_lvl  Termination level of this request. Selects procedures to abort.
+ * @param[out]  p_result  Result of entering the continuous carrier state.
+ */
+static void swi_continuous_carrier(nrf_802154_term_t term_lvl, bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                             = REQ_TYPE_CONTINUOUS_CARRIER;
+    p_slot->data.continuous_carrier.term_lvl = term_lvl;
+    p_slot->data.continuous_carrier.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Notifies the core module that the given buffer is not used anymore and can be freed.
+ *
+ * @param[in]  p_data  Pointer to the buffer to be freed.
+ */
+static void swi_buffer_free(uint8_t * p_data, bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                      = REQ_TYPE_BUFFER_FREE;
+    p_slot->data.buffer_free.p_data   = p_data;
+    p_slot->data.buffer_free.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Notifies the core module that the next higher layer has requested a channel change.
+ */
+static void swi_channel_update(bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                         = REQ_TYPE_CHANNEL_UPDATE;
+    p_slot->data.channel_update.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Notifies the core module that the next higher layer has requested a CCA configuration
+ * change.
+ */
+static void swi_cca_cfg_update(bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                         = REQ_TYPE_CCA_CFG_UPDATE;
+    p_slot->data.cca_cfg_update.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Notifies the core module that the next higher layer requested the RSSI measurement.
+ */
+static void swi_rssi_measure(bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                       = REQ_TYPE_RSSI_MEASURE;
+    p_slot->data.rssi_measure.p_result = p_result;
+
+    req_exit();
+}
+
+/**
+ * @brief Gets the last RSSI measurement result from the core module.
+ */
+static void swi_rssi_measurement_get(int8_t * p_rssi, bool * p_result)
+{
+    nrf_802154_req_data_t * p_slot = req_enter();
+
+    p_slot->type                   = REQ_TYPE_RSSI_GET;
+    p_slot->data.rssi_get.p_rssi   = p_rssi;
+    p_slot->data.rssi_get.p_result = p_result;
+
+    req_exit();
+}
+
 void nrf_802154_request_init(void)
 {
+    nrf_802154_queue_init(&m_requests_queue, m_requests_queue_memory,
+            sizeof(m_requests_queue_memory), sizeof(m_requests_queue_memory[0]));
+
+    nrf_egu_int_enable(SWI_EGU, REQ_INT);
+
     nrf_802154_swi_init();
 }
 
 bool nrf_802154_request_sleep(nrf_802154_term_t term_lvl)
 {
-    REQUEST_FUNCTION(nrf_802154_core_sleep, nrf_802154_swi_sleep, term_lvl)
+    REQUEST_FUNCTION(nrf_802154_core_sleep, swi_sleep, term_lvl)
 }
 
 bool nrf_802154_request_receive(nrf_802154_term_t              term_lvl,
@@ -115,7 +480,7 @@ bool nrf_802154_request_receive(nrf_802154_term_t              term_lvl,
                                 bool                           notify_abort)
 {
     REQUEST_FUNCTION(nrf_802154_core_receive,
-                     nrf_802154_swi_receive,
+                     swi_receive,
                      term_lvl,
                      req_orig,
                      notify_function,
@@ -130,7 +495,7 @@ bool nrf_802154_request_transmit(nrf_802154_term_t              term_lvl,
                                  nrf_802154_notification_func_t notify_function)
 {
     REQUEST_FUNCTION(nrf_802154_core_transmit,
-                     nrf_802154_swi_transmit,
+                     swi_transmit,
                      term_lvl,
                      req_orig,
                      p_data,
@@ -143,45 +508,136 @@ bool nrf_802154_request_energy_detection(nrf_802154_term_t term_lvl,
                                          uint32_t          time_us)
 {
     REQUEST_FUNCTION(nrf_802154_core_energy_detection,
-                     nrf_802154_swi_energy_detection,
+                     swi_energy_detection,
                      term_lvl,
                      time_us)
 }
 
 bool nrf_802154_request_cca(nrf_802154_term_t term_lvl)
 {
-    REQUEST_FUNCTION(nrf_802154_core_cca, nrf_802154_swi_cca, term_lvl)
+    REQUEST_FUNCTION(nrf_802154_core_cca, swi_cca, term_lvl)
 }
 
 bool nrf_802154_request_continuous_carrier(nrf_802154_term_t term_lvl)
 {
-    REQUEST_FUNCTION(nrf_802154_core_continuous_carrier, nrf_802154_swi_continuous_carrier,
+    REQUEST_FUNCTION(nrf_802154_core_continuous_carrier, swi_continuous_carrier,
                      term_lvl)
 }
 
 bool nrf_802154_request_buffer_free(uint8_t * p_data)
 {
-    REQUEST_FUNCTION(nrf_802154_core_notify_buffer_free, nrf_802154_swi_buffer_free, p_data)
+    REQUEST_FUNCTION(nrf_802154_core_notify_buffer_free, swi_buffer_free, p_data)
 }
 
 bool nrf_802154_request_channel_update(void)
 {
-    REQUEST_FUNCTION_NO_ARGS(nrf_802154_core_channel_update, nrf_802154_swi_channel_update)
+    REQUEST_FUNCTION_NO_ARGS(nrf_802154_core_channel_update, swi_channel_update)
 }
 
 bool nrf_802154_request_cca_cfg_update(void)
 {
-    REQUEST_FUNCTION_NO_ARGS(nrf_802154_core_cca_cfg_update, nrf_802154_swi_cca_cfg_update)
+    REQUEST_FUNCTION_NO_ARGS(nrf_802154_core_cca_cfg_update, swi_cca_cfg_update)
 }
 
 bool nrf_802154_request_rssi_measure(void)
 {
-    REQUEST_FUNCTION_NO_ARGS(nrf_802154_core_rssi_measure, nrf_802154_swi_rssi_measure)
+    REQUEST_FUNCTION_NO_ARGS(nrf_802154_core_rssi_measure, swi_rssi_measure)
 }
 
 bool nrf_802154_request_rssi_measurement_get(int8_t * p_rssi)
 {
     REQUEST_FUNCTION(nrf_802154_core_last_rssi_measurement_get,
-                     nrf_802154_swi_rssi_measurement_get,
+                     swi_rssi_measurement_get,
                      p_rssi)
 }
+
+/**@brief Handles REQ_EVENT on SWI_EGU */
+static void irq_handler_req_event(void)
+{
+    while (!nrf_802154_queue_is_empty(&m_requests_queue))
+    {
+        nrf_802154_req_data_t * p_slot =
+                (nrf_802154_req_data_t *)nrf_802154_queue_pop_begin(&m_requests_queue);
+
+        switch (p_slot->type)
+        {
+            case REQ_TYPE_SLEEP:
+                *(p_slot->data.sleep.p_result) =
+                    nrf_802154_core_sleep(p_slot->data.sleep.term_lvl);
+                break;
+
+            case REQ_TYPE_RECEIVE:
+                *(p_slot->data.receive.p_result) =
+                    nrf_802154_core_receive(p_slot->data.receive.term_lvl,
+                                            p_slot->data.receive.req_orig,
+                                            p_slot->data.receive.notif_func,
+                                            p_slot->data.receive.notif_abort);
+                break;
+
+            case REQ_TYPE_TRANSMIT:
+                *(p_slot->data.transmit.p_result) =
+                    nrf_802154_core_transmit(p_slot->data.transmit.term_lvl,
+                                             p_slot->data.transmit.req_orig,
+                                             p_slot->data.transmit.p_data,
+                                             p_slot->data.transmit.cca,
+                                             p_slot->data.transmit.immediate,
+                                             p_slot->data.transmit.notif_func);
+                break;
+
+            case REQ_TYPE_ENERGY_DETECTION:
+                *(p_slot->data.energy_detection.p_result) =
+                    nrf_802154_core_energy_detection(
+                        p_slot->data.energy_detection.term_lvl,
+                        p_slot->data.energy_detection.time_us);
+                break;
+
+            case REQ_TYPE_CCA:
+                *(p_slot->data.cca.p_result) = nrf_802154_core_cca(p_slot->data.cca.term_lvl);
+                break;
+
+            case REQ_TYPE_CONTINUOUS_CARRIER:
+                *(p_slot->data.continuous_carrier.p_result) =
+                    nrf_802154_core_continuous_carrier(
+                        p_slot->data.continuous_carrier.term_lvl);
+                break;
+
+            case REQ_TYPE_BUFFER_FREE:
+                *(p_slot->data.buffer_free.p_result) =
+                    nrf_802154_core_notify_buffer_free(p_slot->data.buffer_free.p_data);
+                break;
+
+            case REQ_TYPE_CHANNEL_UPDATE:
+                *(p_slot->data.channel_update.p_result) = nrf_802154_core_channel_update();
+                break;
+
+            case REQ_TYPE_CCA_CFG_UPDATE:
+                *(p_slot->data.cca_cfg_update.p_result) = nrf_802154_core_cca_cfg_update();
+                break;
+
+            case REQ_TYPE_RSSI_MEASURE:
+                *(p_slot->data.rssi_measure.p_result) = nrf_802154_core_rssi_measure();
+                break;
+
+            case REQ_TYPE_RSSI_GET:
+                *(p_slot->data.rssi_get.p_result) =
+                    nrf_802154_core_last_rssi_measurement_get(p_slot->data.rssi_get.p_rssi);
+                break;
+
+            default:
+                assert(false);
+        }
+
+        nrf_802154_queue_pop_commit(&m_requests_queue);
+    }
+}
+
+void nrf_802154_request_swi_irq_handler(void)
+{
+    if (nrf_egu_event_check(SWI_EGU, REQ_EVENT))
+    {
+        nrf_egu_event_clear(SWI_EGU, REQ_EVENT);
+
+        irq_handler_req_event();
+    }
+}
+
